Use sound device native channel count/freq; use f32 format instead of s16

+++ source_files/edge/snd_data.cc
@@ -66,7 +66,7 @@ void SoundData::Allocate(int samples)
 
     length_ = samples;
 
-    data_ = new int16_t[samples * 2];
+    data_ = new float[samples * 2];
 }
 
 void SoundData::MixSubmerged()
