+++ source_files/edge/snd_gather.h
@@ -38,7 +38,7 @@ class SoundGatherer
     SoundGatherer();
     ~SoundGatherer();
 
-    int16_t *MakeChunk(int max_samples, bool stereo);
+    float *MakeChunk(int max_samples, bool stereo);
     // prepare to add a chunk of sound samples.  Returns a buffer
     // containing the number of samples (* 2 for stereo) which the
     // user can fill up.
